JENKINS-25744 Switch job property to use an optionalBlock
Normal convention for Jenkins plugins which contribute a JobProperty to a jobs
configuration page is to use an optionalBlock as the top level jelly tag.

In a jobs properties section of the configuration page, plugins 'hide' their
configuration entries by using the optionalBlock, this is presented to the user
as a checkbox. This keeps the jobs configuration page quite clean while allowing
the user full control over the properties they use.